move test utility functions to test/utils.py
